+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [FIX] Upgrade flutter_blue_plus version
 * [ADD] Add a method to enable or disable the display backlight
 * [ADD] Add a method to read the firmware version
+* [ADD] Add the construction, help, and camera icons
 
 ## 1.1.1
 * [IMP] Integrate font in text control instruction